Updated matching section
This should be the final change. I have found that m = 1:4 are the only
viable multiples of 2pi if we maintain a constant height. Past that, the
dielectric difference btw one layer and another change too fast. I tried
to solve for eps_r but the quadratic was extraordinarily hairy.